Accept nil player argument.
When technic/technic/machines/HV/quarry.lua calls the bones:can_dig(),
it calls it with a nil player argument. That makes the server crash. I
patched quarry.lua to use the quarry owner's player object, but asl97
rejected the patch, saying:

"The api allows nil player, so it is a problem that the bone mod should
fix.  Also, if the player is offline, minetest.get_player_by_name would
return nil too."